+++ libs/common/src/entities/chat.entity.ts
@@ -6,7 +6,7 @@ import {
   OneToMany,
   PrimaryGeneratedColumn,
 } from 'typeorm';
-import { Message } from '../message/entity/message.entity';
+import { Message } from './message.entity';
 import { User } from '@app/common';
 @Entity('Chat')
 export class Chat {
