+++ libs/common/src/entities/guide.entity.ts
