+++ libs/common/src/entities/message.entity.ts
@@ -1,6 +1,6 @@
 import { Type } from 'class-transformer';
 import { IsBoolean, IsDate, IsNotEmpty, IsString } from 'class-validator';
-import { Chat } from '../../entity/chat.entity';
+import { Chat } from './chat.entity';
 import {
   Column,
   Entity,
