+++ src/chat/message/message.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { MessageService } from './message.service';
 import { MessageController } from './message.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Message } from './entity/message.entity';
+import { Message } from '../../../libs/common/src/entities/message.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Message])],
