+++ src/guide/guide.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { GuideService } from './guide.service';
 import { GuideController } from './guide.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Guide } from './entities/guide.entity';
+import { Guide } from '../../libs/common/src/entities/guide.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Guide])],
